[MOD] Replace app with refactored
