+++ www/index.html
@@ -316,6 +316,6 @@
             </div>
         </main>
         <script src="js/vendor/modernizr-3.12.0.min.js"></script>
-        <script type="module" src="js/main.js"></script>
+        <script type="module" src="js/app.js"></script>
     </body>
 </html>
